Ensure Index but has poblem yet

+++ LiteDB/Utils/StorageUnitHelper.cs
@@ -43,7 +43,7 @@ namespace LiteDB
             var suf = new string[] { "B", "KB", "MB", "GB", "TB" }; //Longs run out around EB
             if (byteCount == 0) return "0" + suf[0];
             var bytes = Math.Abs(byteCount);
-            var place = Convert.ToInt32(Math.Floor(Math.Log(bytes, 1024)));
+            var place = Convert.ToInt64(Math.Floor(Math.Log(bytes, 1024)));
             var num = Math.Round(bytes / Math.Pow(1024, place), 1);
             return (Math.Sign(byteCount) * num).ToString() + suf[place];
         }
